:recycle: rename add event form container
