+++ frontend/src/views/scenes/event/AddScene.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import AddEventContainer from '../../containers/event/AddEventContainer';
+import AddEventFormContainer from '../../containers/event/AddEventFormContainer';
 import NotificationContainer from '../../containers/notification/NotificationContainer';
 import Container from '@material-ui/core/Container';
 import { makeStyles } from '@material-ui/core/styles';
@@ -15,7 +15,7 @@ const AddEventScene: React.FC = () => {
   const classes = useStyles({});
   return (
     <Container className={classes.container}>
-      <AddEventContainer />
+      <AddEventFormContainer />
       <NotificationContainer />
     </Container>
   );
